Move stuff back to wire-protocol
