+++ src/client-configuration.ts
@@ -1,4 +1,4 @@
-import type { QueryValueObject, ValueFormat } from "./wire-protocol";
+import type { ValueFormat } from "./wire-protocol";
 
 /**
  * Configuration for a client. The options provided are be used as
@@ -108,65 +108,6 @@ export interface ClientConfiguration {
    * Enable or disable typechecking of the query before evaluation. If no value
    * is provided, the value of `typechecked` in the database configuration will
    * be used.
-   */
-  typecheck?: boolean;
-}
-
-/**
- * Options for queries. Each query can be made with different options. Settings here
- * take precedence over those in {@link ClientConfiguration}.
- */
-export interface QueryOptions {
-  /** Optional arguments. Variables in the query will be initialized to the
-   * value associated with an argument key.
-   */
-  arguments?: QueryValueObject;
-
-  /**
-   * Determines the encoded format expected for the query `arguments` field, and
-   * the `data` field of a successful response.
-   * Overrides the optional setting on the {@link ClientConfiguration}.
-   */
-  format?: ValueFormat;
-
-  /**
-   * If true, unconditionally run the query as strictly serialized.
-   * This affects read-only transactions. Transactions which write
-   * will always be strictly serialized.
-   * Overrides the optional setting on the {@link ClientConfiguration}.
-   */
-  linearized?: boolean;
-
-  /**
-   * The max number of times to retry the query if contention is encountered.
-   *Overrides the optional setting on the {@link ClientConfiguration}.
-   */
-  max_contention_retries?: number;
-
-  /**
-   * Tags provided back via logging and telemetry.
-   * Overrides the optional setting on the {@link ClientConfiguration}.
-   */
-  query_tags?: Record<string, string>;
-
-  /**
-   * The timeout to use in this query in milliseconds.
-   * Overrides the optional setting on the {@link ClientConfiguration}.
-   */
-  query_timeout_ms?: number;
-
-  /**
-   * A traceparent provided back via logging and telemetry.
-   * Must match format: https://www.w3.org/TR/trace-context/#traceparent-header
-   * Overrides the optional setting on the {@link ClientConfiguration}.
-   */
-  traceparent?: string;
-
-  /**
-   * Enable or disable typechecking of the query before evaluation. If no value
-   * is provided, the value of `typechecked` in the database configuration will
-   * be used.
-   * Overrides the optional setting on the {@link ClientConfiguration}.
    */
   typecheck?: boolean;
 }
